% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELFFR.R
\name{elffr}
\alias{elffr}
\title{Fit longitudinal function-on-function regression models using an efficient 3 step approach}
\usage{
elffr(formula, subj_var, visit_var, data, num_cores = 8, kz = 30, kb = 30, nknots_scalar = 10, nknots_fbps = 5,
                bspline_p = 3, diff_penalty = 2, smooth.cov = TRUE, analytic = TRUE,
                parallel = TRUE, var = TRUE, num_boot = 100, CMA = TRUE,
                family = "gaussian", argvals.functional = NULL, silent = TRUE,
                seed = 1)
}
\arguments{
\item{formula}{Formula object with ff() to specify functional predictors}

\item{subj_var}{Name of subject ID variable in data}

\item{visit_var}{Name of visit/longitudinal variable in data}

\item{data}{Data object generated from the `lfofr_sim_data` function}

\item{num_cores}{If using parallel computing, the number of cores to be used for parallel processing}

\item{kz}{dimension of functional principal component basis for functional predictors}

\item{kb}{number of knots in the spline basis for the functional coefficient}

\item{nknots_scalar}{number of knots for smoothing pointwise estimates for the scalar predictor coefficient}

\item{nknots_fbps}{number of knots for smoothing pointwise estimates for the functional predictor
by the sandwich smoother (refund::fbps). Can be a single number or a list of knots on the domain [0,1] for each dimension}

\item{bspline_p}{degree of B-splines in sandwich smoother}

\item{diff_penalty}{order of differencing penalty in sandwich smoother}

\item{smooth.cov}{logical - whether to smooth covariance matrix in pointwise longitudinal scalar-on-function model}

\item{analytic}{logical - whether to use analytic inference}

\item{parallel}{logical - whether to use parallel computing}

\item{var}{logical - whether to compute variance estimates (else only returns estimated coefficients)}

\item{num_boot}{if using bootstrap inference, number of replicates}

\item{CMA}{logical - whether to compute correlation and multiplicity adjusted (CMA) confidence bands}

\item{family}{character - set to "gaussian" for normally-distributed outcomes}

\item{argvals.functional}{list specifying numeric arguments of functional predictors
in the same order as the ff() terms supplied to `formula`
Used for adjusting scale of estimates. If NA, defaults to equally spaced interval 1,...,U.
Note, ELFFR only works for data on regularly spaced grids.}

\item{seed}{seed used for drawing samples in bootstrap inference}
}
\value{
For the case of variance estimation with analytic inference:
\item{`scalar.est`}{estimated coefficient curves for scalar predictors}
\item{`functional.est`}{estimated coefficient surfaces for functional predictors}
\item{`var.scalar`}{variance estimates for scalar predictor coefficient curves}
\item{`var.functional`}{variance estimates for functional predictor coefficient surfaces}
\item{`time`}{time to fit in seconds}
\item{`CMA_quantiles`}{quantiles for CMA confidence regions}
}
\description{
`elffr` (efficient longitudinal function-on-function regression) fits longitudinal function-on-function regression models by (1) fitting
longitudinal scalar-on-function models with a random intercept at each index of the response domain, (2) smoothing pointwise
fixed effect estimates by univariate/bivariate smoothers and (3) constructing confidence regions for coefficient estimates
by either an analytic approach in the case of Gaussian responses or by a bootstrap procedure for general response distributions.
}
