% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gamma_smoothers.R
\name{get_gamma_smoothers}
\alias{get_gamma_smoothers}
\title{Helper function to obtain sandwich smoother matrices
for analytic inference}
\usage{
get_gamma_smoothers(
  data,
  knots,
  p,
  m,
  subj = NULL,
  covariates = NULL,
  knots.option = "equally-spaced",
  periodicity = c(FALSE, FALSE),
  lambda = NULL,
  selection = "GCV",
  search.grid = T,
  search.length = 100,
  method = "L-BFGS-B",
  lower = -20,
  upper = 20,
  control = NULL
)
}
\description{
Helper function to obtain sandwich smoother matrices
for analytic inference
}
