% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpfr_boot.R
\name{lpfr_boot}
\alias{lpfr_boot}
\title{Helper function to run bootstrapped marginal LPFR models}
\usage{
lpfr_boot(
  s_l,
  b,
  boot_dat,
  temp_dir,
  kz,
  kb,
  smooth.cov,
  family,
  scalar_varnames,
  ff_varnames,
  Y_label
)
}
\description{
Helper function to run bootstrapped marginal LPFR models
}
