% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffrSim_modified.R
\name{pffrSim_modified}
\alias{pffrSim_modified}
\title{Simulate example data for lfofr - adapted from pffrSim in refund}
\usage{
pffrSim_modified(
  func_predictor,
  scenario = "all",
  n = 100,
  nxgrid = 40,
  nygrid = 60,
  SNR = 10,
  propmissing = 0,
  limits = NULL
)
}
\arguments{
\item{scenario}{see Description}

\item{n}{number of observations}

\item{nxgrid}{number of evaluation points of functional covariates}

\item{nygrid}{number of evaluation points of the functional response}

\item{SNR}{the signal-to-noise ratio for the generated data: empirical
variance of the additive predictor divided by variance of the errors.}

\item{propmissing}{proportion of missing data in the response, default = 0.
See Details.}

\item{limits}{a function that defines an integration range, see
\code{\link{ff}}}
}
\value{
a named list with the simulated data, and the true components of the
  predictor etc as attributes.
}
\description{
Simulate example data for lfofr - adapted from pffrSim in refund
}
