% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_boot.R
\name{run_boot}
\alias{run_boot}
\title{Helper function to fit bootstrapped ELFFR models}
\usage{
run_boot(
  formula,
  data,
  subj_var,
  visit_var,
  seed,
  p,
  argvals,
  U.list,
  num_boot = 50,
  splines = "ps",
  nknots_scalar = 5,
  nknots_fbps = 3,
  bspline_p = 3,
  diff_penalty = 2,
  scenario = "default",
  num_cores = 8,
  kz = 30,
  kb = 30,
  smooth.cov = TRUE,
  family = "gaussian",
  cl = NA
)
}
\description{
Helper function to fit bootstrapped ELFFR models
}
