% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_elffr_data.R
\name{sim_elffr_data}
\alias{sim_elffr_data}
\title{Generate data to be used for longitudinal FoFR}
\usage{
sim_elffr_data(
  func_predictors,
  family = "gaussian",
  I = 100,
  J = 5,
  L = 100,
  U = 100,
  SNR_B = 1,
  SNR_sigma = 1,
  seed = 1,
  psi2 = FALSE
)
}
\arguments{
\item{func_predictors}{list of strings of mathematical expressions with s and u as arguments for the response and predictor domain
defining the true functional predictor coefficient functions}

\item{family}{family of response function: "gaussian", "binomial", or "poisson"}

\item{I}{number of subjects}

\item{J}{average number of visits for each subject}

\item{L}{dimension of response domain}

\item{U}{dimension of predictor domains - list if more than 1 functional predictor, else a number}

\item{SNR_B}{relative importance of random effects}

\item{SNR_sigma}{signal-to-noise ratio}

\item{seed}{integer seed for generating random data}

\item{psi2}{whether to include subject-visit specific random effects (otherwise only subject-specific)}
}
\value{
Returns a dataframe with (1) ID - integer for subject identification, (2) visit - visit number for the
i'th subject, (3) X. scalar predictors (4) W. - functional predictors, (5) Y - response, and
(6) eta - true response / linear predictor value
}
\description{
Generate data to be used for longitudinal FoFR
}
